feat: add new class
